Move notifications endpoint to separate REST controller
